refactor: Colocate every styles together
