+++ packages/cozy-sharing/src/components/CozyPassFingerprintDialogContent.jsx
@@ -7,7 +7,7 @@ import Paper from 'cozy-ui/transpiled/react/Paper'
 import Typography from 'cozy-ui/transpiled/react/Typography'
 import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
 
-import styles from './CozyPassFingerprintDialogContent.styl'
+import styles from '../styles/CozyPassFingerprintDialogContent.styl'
 
 /**
  * CozyPassFingerprintDialogContent can be use as a content for ShareModal when
