+++ packages/cozy-sharing/src/components/Recipient/OwnerIdentity.jsx
@@ -3,8 +3,8 @@ import React from 'react'
 import Avatar from 'cozy-ui/transpiled/react/Avatar'
 
 import Identity from './Identity'
-import styles from './recipient.styl'
 import { getDisplayName, getInitials } from '../../models'
+import styles from '../../styles/recipient.styl'
 
 const OwnerIdentity = ({ url, size, ...rest }) => (
   <div className={styles['avatar']}>
