+++ packages/cozy-sharing/src/components/SharedStatus.jsx
@@ -5,9 +5,9 @@ import palette from 'cozy-ui/transpiled/react/palette'
 import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
 
 import { SharingTooltip, TooltipRecipientList } from './Tooltip'
-import styles from './status.styl'
 import LinkIcon from '../../assets/icons/icon-link.svg'
 import { getDisplayName } from '../models'
+import styles from '../styles/status.styl'
 
 export const SharedStatus = ({ className, docId, recipients, link }) => {
   const { t } = useI18n()
