+++ packages/cozy-sharing/src/components/Tooltip.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import ReactTooltip from 'react-tooltip'
 
-import styles from './tooltip.styl'
+import styles from '../styles/tooltip.styl'
 
 export const SharingTooltip = props => (
   <ReactTooltip
