+++ packages/cozy-sharing/src/styles/CozyPassFingerprintDialogContent.styl
