+++ packages/cozy-sharing/src/styles/actionmenu.styl
