+++ packages/cozy-sharing/src/styles/autosuggest.styl
