+++ packages/cozy-sharing/src/styles/badge.styl
