+++ packages/cozy-sharing/src/styles/button.styl
