+++ packages/cozy-sharing/src/styles/publicBanner.styl
@@ -6,9 +6,9 @@
     img
         @extend $avatar
         @extend $circle
-        width rem(16)	  
-        height rem(16)	    
-        min-width rem(16)	
-        min-height rem(16)	
+        width rem(16)
+        height rem(16)
+        min-width rem(16)
+        min-height rem(16)
         font-size rem(7)
         vertical-align text-bottom
