+++ packages/cozy-sharing/src/styles/recipient.styl
