+++ packages/cozy-sharing/src/styles/share.styl
