+++ packages/cozy-sharing/src/styles/status.styl
