+++ packages/cozy-sharing/src/styles/tooltip.styl
